Renames files.
