+++ bin/con_man_gui.rb
@@ -2,12 +2,12 @@
 
 $LOAD_PATH.unshift File.dirname(__FILE__) + '/../lib'
 
-require 'add_contact'
+require 'add_contact_window'
 require 'sql_contacts_repository'
 require 'qt'
 
 contacts_repository = SQLContactsRepository.new
 app = Qt::Application.new (ARGV)
-window = AddContact.new(contacts_repository)
+window = AddContactWindow.new(contacts_repository)
 window.show
 app.exec
