+++ lib/contacts.rb
