+++ lib/sql_contacts_repository.rb
@@ -1,14 +1,14 @@
-require 'contacts_db'
+require 'contacts'
 
 class SQLContactsRepository
 
   def add(contact)
     Contacts.create(:first_name => contact.first_name,
-                    :last_name => contact.last_name,
-                    :dob => contact.dob,
-                    :telephone => contact.telephone,
-                    :email => contact.email,
-                    :address => contact.address)
+    :last_name => contact.last_name,
+    :dob => contact.dob,
+    :telephone => contact.telephone,
+    :email => contact.email,
+    :address => contact.address) 
   end
 
   def found(contact)
